2003-06-29  Morten Brix Pedersen  <morten@wtf.dk>
        * NickList.h: Change 'int' to 'unsigned short'. It's a smaller data
          type, dont know how much it will mean in the real world.
        * ServerConnection.cpp: Fix problem with having duplicates in the user
          list when reconnecting to a server. The disconnected() signal was
          never sent to the frontend.

+++ lostirc/src/gtkmm/NickList.h
@@ -47,7 +47,7 @@ private:
     {
         Gtk::TreeModelColumn<Glib::ustring> status;
         Gtk::TreeModelColumn<Glib::ustring> nick;
-        Gtk::TreeModelColumn<int> priority;
+        Gtk::TreeModelColumn<unsigned short> priority;
 
         ModelColumns() { add(status); add(nick); add(priority); }
     };
